requirements.txt
numpy
pandas
PyYAML
scikit_learn
scipy
torch
torch_geometric
tqdm
